Makes the Block class into a base class with different implementations (DataBlock, MetaBlock, IndexBlock, etc).

This change can allow different implementations of these base classes for use for different iterator implementations.  For example, it is possible to create different data block and iterator implementations that un-"delta encode" the values, which may yield performance improvements under certain circumstances.

TODO: There are a few tests that fail now because this change exposes #8673.

TODO: Do a better job of storing/using "previous" values in the DataIterator.