@mrambacher 2021-09-28
> Still Work In Progress?

This is "ready for review" now.  This PR is phase 1 in a multi-step process to allow different implementations of Block (specifically DataBlock) to test performance.  Phase 2 will be allowing different implementations of DataBlockIter and Phase 3 will be testing different implementations to see if we can improve performance.