@ajkr 2021-09-29
> I would not be surprised if there is some subtle regression in here that is not caught by unit tests. Please run crash test locally for some hours before shipping.

What about performance tests?