@mrambacher 2021-09-29
> > I would not be surprised if there is some subtle regression in here that is not caught by unit tests. Please run crash test locally for some hours before shipping.
> 
> What about performance tests?

I have executed the performance tests from the Wiki page (with also revrange and revrangewhilewriting tests) and so no regressions.  I can post the results/comparison in the PR if it helps.

I have not executed stress tests but will do so shortly.