@ajkr 2021-09-29
Here are the commands I'm using:

```
$ TEST_TMPDIR=/dev/shm ./db_bench -benchmarks=filluniquerandom,compact -target_file_size_base=1048576 -compression_type=none
$ make clean && EXTRA_LDFLAGS="-fprofile-generate" EXTRA_CXXFLAGS="-fprofile-generate" V=1 USE_LTO=1 DEBUG_LEVEL=0 OPTIMIZE_LEVEL=-O3 make -j48 db_bench
$ TEST_TMPDIR=/dev/shm ./db_bench -readonly=1 -use_existing_db=1 -benchmarks=seekrandom -seek_nexts=1000 -target_file_size_base=1048576 -cache_size=0 -duration=30
$ find . -name '*.o' | xargs rm -f
$ EXTRA_LDFLAGS="-fprofile-use -fprofile-correction" EXTRA_CXXFLAGS="-Wno-error=missing-profile -fprofile-use -fprofile-correction" V=1 USE_LTO=1 DEBUG_LEVEL=0 OPTIMIZE_LEVEL=-O3 make -j48 db_bench
$ TEST_TMPDIR=/dev/shm ./db_bench -readonly=1 -use_existing_db=1 -benchmarks=seekrandom -seek_nexts=1000 -target_file_size_base=1048576 -cache_size=0 -duration=30
```

Here are the results.

Results before 8690: 962.0MB/s, *930.0MB/s*, 916.5MB/s
Results after 8690: *838.5MB/s*, 850.6MB/s, 822.5MB/s

I will do a second build and run from scratch to make sure no mistakes.