@ajkr 2021-09-30
Here are some commands I found to exaggerate the regression up to 15% while using simpler build options:

```
$ TEST_TMPDIR=/dev/shm ./db_bench -benchmarks=filluniquerandom,compact -target_file_size_base=1048576 -compression_type=none
$ make clean && V=1 DEBUG_LEVEL=0 OPTIMIZE_LEVEL=-O3 make -j48 db_bench
$ TEST_TMPDIR=/dev/shm ./db_bench -mmap_read=1 -verify_checksum=false -use_existing_db=1 -benchmarks=seekrandom -seek_nexts=1000 -target_file_size_base=1048576 -cache_size=0 -duration=30
```